+++ example/lib/src/1/home_page.dart
@@ -1,4 +1,3 @@
-
 import 'package:flutter/material.dart';
 
 import 'package:set_state/set_state.dart';
@@ -7,21 +6,23 @@ import 'package:example/src/1/home_bloc.dart';
 
 import 'package:example/src/2/second_page.dart';
 
-
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key}) : _state = _MyHomePageState(), super(key: key);
-  final _MyHomePageState _state;
+  factory MyHomePage({Key key}) => _this ??= MyHomePage._(key: key);
+  MyHomePage._({Key key}) : super(key: key);
+  static MyHomePage _this;
 
   @override
-  _MyHomePageState createState() => _state;
+  _MyHomePageState createState() => _MyHomePageState();
 
-  /// Supply a means to access its bloc
-  void onPressed() => _state.onPressed();
+  // Supply a means to access its bloc
+  void onPressed() {
+    final _MyHomePageState state = SetState.to<_MyHomePageState>();
+    state.onPressed();
+  }
 }
 
 /// The home page for this app.
 class _MyHomePageState extends SetState<MyHomePage> {
-  //
   _MyHomePageState() : super() {
     bloc = HomePageBloc<_MyHomePageState>();
   }
@@ -39,40 +40,39 @@ class _MyHomePageState extends SetState<MyHomePage> {
 
   @override
   Widget build(BuildContext context) => Scaffold(
-    appBar: AppBar(
-      title: const Text('Home Page'),
-    ),
-    body: Center(
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: <Widget>[
-          const Text(
-            'You have pushed the button this many times:',
+        appBar: AppBar(
+          title: const Text('Home Page'),
+        ),
+        body: Center(
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              const Text(
+                'You have pushed the button this many times:',
+              ),
+              Text(
+                '${bloc.data}',
+                style: Theme.of(context).textTheme.headline4,
+              ),
+            ],
           ),
-          Text(
-            '${bloc.data}',
-            style: Theme.of(context).textTheme.headline4,
+        ),
+        floatingActionButton: FloatingActionButton(
+          onPressed: onPressed,
+          child: Icon(Icons.add),
+        ),
+        persistentFooterButtons: <Widget>[
+          RaisedButton(
+            child: const Text(
+              'Second Page',
+            ),
+            onPressed: () {
+              Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                      builder: (BuildContext context) => SecondPage()));
+            },
           ),
         ],
-      ),
-    ),
-    floatingActionButton: FloatingActionButton(
-      onPressed: onPressed,
-      child: Icon(Icons.add),
-    ),
-    persistentFooterButtons: <Widget>[
-      RaisedButton(
-        child: const Text(
-          'Second Page',
-        ),
-        onPressed: () {
-          Navigator.push(
-              context,
-              MaterialPageRoute(
-                  builder: (BuildContext context) => SecondPage()));
-        },
-      ),
-    ],
-  );
+      );
 }
-